cleanup app code
