+++ hello_vsi/source/application/gui/gui.h
@@ -19,8 +19,8 @@
 #ifndef __GUI_H__
 #define __GUI_H__
 
-void setup_gui(void* argument);
+void init_gui(void* argument);
 
-void loop_gui(void);
+void run_gui(void);
 
 #endif //__GUI_H__
